Cleaned up the code so now its a little more readable. Added start and end time for the visual report. now i can move to the next step that is the tagger
