+++ WPFHook/ViewModels/Hooks/WindowHook.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Text;
+using WPFHook.ViewModels.BackgroundLogic;
 
 namespace WPFHook.ViewModels
 {
